Debugged tank drive code as well as drive base code in general in order to make it function.

+++ Quasics2021Code/MaeTester/src/main/include/Constants.h
@@ -79,5 +79,5 @@ namespace OIConstants {
 }  // namespace OIConstants
 
 namespace DrivebaseConstants{
-  constexpr int powerScaling = .6;
+  constexpr double powerScaling = .6;
 }
